feat: add widgets
