+++ android/app/src/main/res/values/colors.xml
@@ -11,4 +11,8 @@
     <!--End Widget-->
 
     <color name="ic_launcher_background">#F9F0F0</color>
+    <color name="light_blue_50">#FFE1F5FE</color>
+    <color name="light_blue_200">#FF81D4FA</color>
+    <color name="light_blue_600">#FF039BE5</color>
+    <color name="light_blue_900">#FF01579B</color>
 </resources>
